Rename Locales to Locale
