+++ includes/Admin/Settings_Screens/Messenger.php
@@ -14,7 +14,7 @@ defined( 'ABSPATH' ) or exit;
 
 use SkyVerge\WooCommerce\Facebook\Admin;
 use SkyVerge\WooCommerce\Facebook\API\FBE\Configuration;
-use SkyVerge\WooCommerce\Facebook\Locales;
+use SkyVerge\WooCommerce\Facebook\Locale;
 use SkyVerge\WooCommerce\PluginFramework\v5_9_0 as Framework;
 
 /**
@@ -76,7 +76,7 @@ class Messenger extends Admin\Abstract_Settings_Screen {
 		}
 
 		$configured_locale = $this->remote_configuration->get_default_locale();
-		$supported_locales = Locales::get_supported_locales();
+		$supported_locales = Locale::get_supported_locales();
 
 		if ( ! empty( $supported_locales[ $configured_locale ] ) ) {
 			$configured_locale = $supported_locales[ $configured_locale ];
